requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
Flask-Migrate>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
pandas>=2.0
openpyxl>=3.1